refactor: pricefeed (clob ws) its own service
